+++ crates/fuel-core/src/service/genesis/importer.rs
@@ -30,6 +30,11 @@ use crate::{
         coins::OwnedCoins,
         contracts::ContractsInfo,
         messages::OwnedMessageIds,
+        old::{
+            OldFuelBlockConsensus,
+            OldFuelBlocks,
+            OldTransactions,
+        },
         transactions::{
             OwnedTransactions,
             TransactionStatuses,
@@ -52,7 +57,9 @@ use fuel_core_storage::{
         ContractsLatestUtxo,
         ContractsRawCode,
         ContractsState,
+        FuelBlocks,
         Messages,
+        SealedBlockConsensus,
         Transactions,
     },
 };
@@ -110,13 +117,19 @@ impl SnapshotImporter {
         self.spawn_worker_on_chain::<ContractsLatestUtxo>()?;
         self.spawn_worker_on_chain::<ContractsState>()?;
         self.spawn_worker_on_chain::<ContractsAssets>()?;
-        self.spawn_worker_on_chain::<Transactions>()?;
 
         self.spawn_worker_off_chain::<TransactionStatuses, TransactionStatuses>()?;
         self.spawn_worker_off_chain::<OwnedTransactions, OwnedTransactions>()?;
         self.spawn_worker_off_chain::<Messages, OwnedMessageIds>()?;
         self.spawn_worker_off_chain::<Coins, OwnedCoins>()?;
+        self.spawn_worker_off_chain::<FuelBlocks, OldFuelBlocks>()?;
+        self.spawn_worker_off_chain::<Transactions, OldTransactions>()?;
+        self.spawn_worker_off_chain::<SealedBlockConsensus, OldFuelBlockConsensus>()?;
         self.spawn_worker_off_chain::<Transactions, ContractsInfo>()?;
+        self.spawn_worker_off_chain::<OldTransactions, ContractsInfo>()?;
+        self.spawn_worker_off_chain::<OldFuelBlocks, OldFuelBlocks>()?;
+        self.spawn_worker_off_chain::<OldFuelBlockConsensus, OldFuelBlockConsensus>()?;
+        self.spawn_worker_off_chain::<OldTransactions, OldTransactions>()?;
 
         self.task_manager.wait().await?;
 
@@ -128,7 +141,7 @@ impl SnapshotImporter {
         TableBeingWritten: TableWithBlueprint + 'static + Send,
         TableEntry<TableBeingWritten>: serde::de::DeserializeOwned + Send,
         StateConfig: AsTable<TableBeingWritten>,
-        Handler<TableBeingWritten>:
+        Handler<TableBeingWritten, TableBeingWritten>:
             ImportTable<TableInSnapshot = TableBeingWritten, DbDesc = OnChain>,
     {
         let groups = self.snapshot_reader.read::<TableBeingWritten>()?;
@@ -138,7 +151,9 @@ impl SnapshotImporter {
         let da_block_height = self.genesis_block.header().da_height;
         let db = self.db.on_chain().clone();
 
-        let progress_reporter = self.progress_reporter::<TableBeingWritten>(num_groups);
+        let progress_name = migration_name::<TableBeingWritten, TableBeingWritten>();
+
+        let progress_reporter = self.progress_reporter(progress_name, num_groups);
 
         self.task_manager.spawn(move |token| {
             let task = ImportTask::new(
@@ -158,10 +173,10 @@ impl SnapshotImporter {
         &mut self,
     ) -> anyhow::Result<()>
     where
-        TableInSnapshot: TableWithBlueprint + 'static,
+        TableInSnapshot: TableWithBlueprint + Send + 'static,
         TableEntry<TableInSnapshot>: serde::de::DeserializeOwned + Send,
         StateConfig: AsTable<TableInSnapshot>,
-        Handler<TableBeingWritten>:
+        Handler<TableBeingWritten, TableInSnapshot>:
             ImportTable<TableInSnapshot = TableInSnapshot, DbDesc = OffChain>,
         TableBeingWritten: TableWithBlueprint + Send + 'static,
     {
@@ -172,7 +187,10 @@ impl SnapshotImporter {
 
         let db = self.db.off_chain().clone();
 
-        let progress_reporter = self.progress_reporter::<TableBeingWritten>(num_groups);
+        let progress_reporter = self.progress_reporter(
+            migration_name::<TableInSnapshot, TableBeingWritten>(),
+            num_groups,
+        );
 
         self.task_manager.spawn(move |token| {
             let task = ImportTask::new(
@@ -200,18 +218,15 @@ impl SnapshotImporter {
         std::io::stderr().is_terminal() && !cfg!(test)
     }
 
-    fn progress_reporter<T>(&self, num_groups: usize) -> ProgressReporter
-    where
-        T: TableWithBlueprint,
-    {
+    fn progress_reporter(&self, name: String, num_groups: usize) -> ProgressReporter {
         let target = if Self::should_display_bars() {
-            Target::Cli(T::column().name())
+            Target::Cli(name)
         } else {
             let span = tracing::span!(
                 parent: &self.tracing_span,
                 Level::INFO,
                 "task",
-                table = T::column().name()
+                table = name
             );
             Target::Logs(span)
         };
@@ -222,18 +237,32 @@ impl SnapshotImporter {
 }
 
 #[derive(Debug, Clone, Copy)]
-pub struct Handler<T> {
+pub struct Handler<TableBeingWritten, TableInSnapshot> {
     pub block_height: BlockHeight,
     pub da_block_height: DaBlockHeight,
-    pub phaton_data: PhantomData<T>,
+    _table_being_written: PhantomData<TableBeingWritten>,
+    _table_in_snapshot: PhantomData<TableInSnapshot>,
 }
 
-impl<T> Handler<T> {
+impl<A, B> Handler<A, B> {
     pub fn new(block_height: BlockHeight, da_block_height: DaBlockHeight) -> Self {
         Self {
             block_height,
             da_block_height,
-            phaton_data: PhantomData,
+            _table_being_written: PhantomData,
+            _table_in_snapshot: PhantomData,
         }
     }
+}
+
+fn migration_name<TableInSnapshot, TableBeingWritten>() -> String
+where
+    TableInSnapshot: TableWithBlueprint,
+    TableBeingWritten: TableWithBlueprint,
+{
+    format!(
+        "{} -> {}",
+        TableInSnapshot::column().name(),
+        TableBeingWritten::column().name()
+    )
 }