+++ crates/fuel-core/src/service/genesis/importer/off_chain.rs
@@ -11,6 +11,11 @@ use crate::{
             coins::OwnedCoins,
             contracts::ContractsInfo,
             messages::OwnedMessageIds,
+            old::{
+                OldFuelBlockConsensus,
+                OldFuelBlocks,
+                OldTransactions,
+            },
             transactions::{
                 OwnedTransactions,
                 TransactionStatuses,
@@ -23,7 +28,9 @@ use fuel_core_chain_config::TableEntry;
 use fuel_core_storage::{
     tables::{
         Coins,
+        FuelBlocks,
         Messages,
+        SealedBlockConsensus,
         Transactions,
     },
     transactional::StorageTransaction,
@@ -36,7 +43,7 @@ use super::{
     Handler,
 };
 
-impl ImportTable for Handler<TransactionStatuses> {
+impl ImportTable for Handler<TransactionStatuses, TransactionStatuses> {
     type TableInSnapshot = TransactionStatuses;
     type TableBeingWritten = TransactionStatuses;
     type DbDesc = OffChain;
@@ -54,7 +61,7 @@ impl ImportTable for Handler<TransactionStatuses> {
     }
 }
 
-impl ImportTable for Handler<FuelBlockIdsToHeights> {
+impl ImportTable for Handler<FuelBlockIdsToHeights, FuelBlockIdsToHeights> {
     type TableInSnapshot = FuelBlockIdsToHeights;
     type TableBeingWritten = FuelBlockIdsToHeights;
     type DbDesc = OffChain;
@@ -72,7 +79,7 @@ impl ImportTable for Handler<FuelBlockIdsToHeights> {
     }
 }
 
-impl ImportTable for Handler<OwnedTransactions> {
+impl ImportTable for Handler<OwnedTransactions, OwnedTransactions> {
     type TableInSnapshot = OwnedTransactions;
     type TableBeingWritten = OwnedTransactions;
     type DbDesc = OffChain;
@@ -90,7 +97,7 @@ impl ImportTable for Handler<OwnedTransactions> {
     }
 }
 
-impl ImportTable for Handler<OwnedMessageIds> {
+impl ImportTable for Handler<OwnedMessageIds, Messages> {
     type TableInSnapshot = Messages;
     type TableBeingWritten = OwnedMessageIds;
     type DbDesc = OffChain;
@@ -108,7 +115,7 @@ impl ImportTable for Handler<OwnedMessageIds> {
     }
 }
 
-impl ImportTable for Handler<OwnedCoins> {
+impl ImportTable for Handler<OwnedCoins, Coins> {
     type TableInSnapshot = Coins;
     type TableBeingWritten = OwnedCoins;
     type DbDesc = OffChain;
@@ -126,7 +133,7 @@ impl ImportTable for Handler<OwnedCoins> {
     }
 }
 
-impl ImportTable for Handler<ContractsInfo> {
+impl ImportTable for Handler<ContractsInfo, Transactions> {
     type TableInSnapshot = Transactions;
     type TableBeingWritten = ContractsInfo;
     type DbDesc = OffChain;
@@ -138,6 +145,130 @@ impl ImportTable for Handler<ContractsInfo> {
     ) -> anyhow::Result<()> {
         let transactions = group.iter().map(|TableEntry { value, .. }| value);
         worker_service::process_transactions(transactions, tx)?;
+        Ok(())
+    }
+}
+
+impl ImportTable for Handler<ContractsInfo, OldTransactions> {
+    type TableInSnapshot = OldTransactions;
+    type TableBeingWritten = ContractsInfo;
+    type DbDesc = OffChain;
+
+    fn process(
+        &mut self,
+        group: Vec<TableEntry<Self::TableInSnapshot>>,
+        tx: &mut StorageTransaction<&mut Database<Self::DbDesc>>,
+    ) -> anyhow::Result<()> {
+        let transactions = group.iter().map(|TableEntry { value, .. }| value);
+        worker_service::process_transactions(transactions, tx)?;
+        Ok(())
+    }
+}
+
+impl ImportTable for Handler<OldFuelBlocks, FuelBlocks> {
+    type TableInSnapshot = FuelBlocks;
+    type TableBeingWritten = OldFuelBlocks;
+    type DbDesc = OffChain;
+
+    fn process(
+        &mut self,
+        group: Vec<TableEntry<Self::TableInSnapshot>>,
+        tx: &mut StorageTransaction<&mut Database<Self::DbDesc>>,
+    ) -> anyhow::Result<()> {
+        let blocks = group
+            .iter()
+            .map(|TableEntry { key, value, .. }| (key, value));
+        worker_service::copy_to_old_blocks(blocks, tx)?;
+        Ok(())
+    }
+}
+
+impl ImportTable for Handler<OldFuelBlocks, OldFuelBlocks> {
+    type TableInSnapshot = OldFuelBlocks;
+    type TableBeingWritten = OldFuelBlocks;
+    type DbDesc = OffChain;
+
+    fn process(
+        &mut self,
+        group: Vec<TableEntry<Self::TableInSnapshot>>,
+        tx: &mut StorageTransaction<&mut Database<Self::DbDesc>>,
+    ) -> anyhow::Result<()> {
+        let blocks = group
+            .iter()
+            .map(|TableEntry { key, value, .. }| (key, value));
+        worker_service::copy_to_old_blocks(blocks, tx)?;
+        Ok(())
+    }
+}
+
+impl ImportTable for Handler<OldFuelBlockConsensus, SealedBlockConsensus> {
+    type TableInSnapshot = SealedBlockConsensus;
+    type TableBeingWritten = OldFuelBlockConsensus;
+    type DbDesc = OffChain;
+
+    fn process(
+        &mut self,
+        group: Vec<TableEntry<Self::TableInSnapshot>>,
+        tx: &mut StorageTransaction<&mut Database<Self::DbDesc>>,
+    ) -> anyhow::Result<()> {
+        let blocks = group
+            .iter()
+            .map(|TableEntry { key, value, .. }| (key, value));
+        worker_service::copy_to_old_block_consensus(blocks, tx)?;
+        Ok(())
+    }
+}
+
+impl ImportTable for Handler<OldFuelBlockConsensus, OldFuelBlockConsensus> {
+    type TableInSnapshot = OldFuelBlockConsensus;
+    type TableBeingWritten = OldFuelBlockConsensus;
+    type DbDesc = OffChain;
+
+    fn process(
+        &mut self,
+        group: Vec<TableEntry<Self::TableInSnapshot>>,
+        tx: &mut StorageTransaction<&mut Database<Self::DbDesc>>,
+    ) -> anyhow::Result<()> {
+        let blocks = group
+            .iter()
+            .map(|TableEntry { key, value, .. }| (key, value));
+        worker_service::copy_to_old_block_consensus(blocks, tx)?;
+        Ok(())
+    }
+}
+
+impl ImportTable for Handler<OldTransactions, Transactions> {
+    type TableInSnapshot = Transactions;
+    type TableBeingWritten = OldTransactions;
+    type DbDesc = OffChain;
+
+    fn process(
+        &mut self,
+        group: Vec<TableEntry<Self::TableInSnapshot>>,
+        tx: &mut StorageTransaction<&mut Database<Self::DbDesc>>,
+    ) -> anyhow::Result<()> {
+        let transactions = group
+            .iter()
+            .map(|TableEntry { key, value, .. }| (key, value));
+        worker_service::copy_to_old_transactions(transactions, tx)?;
+        Ok(())
+    }
+}
+
+impl ImportTable for Handler<OldTransactions, OldTransactions> {
+    type TableInSnapshot = OldTransactions;
+    type TableBeingWritten = OldTransactions;
+    type DbDesc = OffChain;
+
+    fn process(
+        &mut self,
+        group: Vec<TableEntry<Self::TableInSnapshot>>,
+        tx: &mut StorageTransaction<&mut Database<Self::DbDesc>>,
+    ) -> anyhow::Result<()> {
+        let transactions = group
+            .iter()
+            .map(|TableEntry { key, value, .. }| (key, value));
+        worker_service::copy_to_old_transactions(transactions, tx)?;
         Ok(())
     }
 }